Add preference for citekey label in extra field
This is still a bit quirky.  If the preference is changed, the cache
must be cleared manually before Better BibTeX figures out either (1)
that an item with the old label no longer has a fixed citation or (2) an
item with the new label now has a fixed citation.  Still, the preference
is unlikely to be changed frequently, so it's usable.